requirements.txt
pillow==9.2.0
dearpygui==1.6.1
